fix: packaging problem
As this project uses EJS without using express.js so the template
need to be read from the desk, render, then write to the desk
and finally, open it using Electron(chromium process).
Using electron-builder to package the app, electron-builder packages
all source code in archive called app.asar, in runtime this archive is
read-only and if I try to write to it, it shows unexpected behavior.
So when trying to write the template after rendering it, it was written
to the archive app.asar which is not allowed so, the option "asarUnpack"
in package.json tells electron-builder what not to add to the app.asar
archive, after specifying the directory where the write happens the
bug is fixed.